Update 12/10 7:07 PM
Made some more changes to the Final Project Script, as well as added
Dead Weight as a variable parameter to the matrix set up.  Fixed some
things in Optimization.  Next steps include getting Train Optimization
function to run (I tried running it but it was taking too long for me
to wait - options to fix this).  Then displaying the data we get from
the optimization function.  I am probably done for the day with this